Small tidyups on Cycler. Add TapTempo class - needs tweaks to make offset intuitive.

+++ src/cycler.h
@@ -24,7 +24,7 @@ class Cycler {
         };
 
         Cycler(
-            unsigned long period=1000,
+            uint16_t period=1000,
             float min=0.0,
             float max=255.0,
             mode_t cycle_mode=STATIC,
@@ -37,16 +37,16 @@ class Cycler {
         void set_min(float min);
         void set_max(float max);
         void set_duty(float duty);
-        void set_period(unsigned long period, bool maintain_progress=false);
+        void set_period(uint16_t period, bool maintain_progress=false);
         void start_period_now();
-        void set_offset(unsigned long offset);
+        void set_offset(uint16_t offset);
 
     private:
         // How long it takes to get to the same point in the cycle in milliseconds
-        unsigned long _period;
+        uint16_t _period;
 
         // The offset from 0 for when the period starts
-        unsigned long _milli_offset;
+        uint16_t _milli_offset;
 
         // The last updates progress through the period normalised between 0 and 1
         float _last_normalised_progress;
@@ -66,7 +66,7 @@ class Cycler {
         // When the cycler was last updated
         unsigned long _last_update_time;
 
-        // The duty or ratio of on to off when in SQUARE cycle mode
+        // The duty or ratio of on to off when in SQUARE or TRIANGLE scycle mode
         float _duty;
 
         // Whetehr the callbacks have been invalidated by changing duty or period
